almost there...

+++ versioning.py
@@ -13,122 +13,184 @@
 # pylint: disable=E0401,E1101,E0611
 
 import os
+import sys
 from os.path import join, exists, isfile, isdir, dirname, basename
-try:
-    from configparser import ConfigParser
-except:
-    from ConfigParser import ConfigParser
 
-def find_parent_containing(name, path=None, check='exists'):
-    """ Return the nearest directory in the parent dirs of path,
-        that contains name, or None if no such parent dir exists.
-        The check can be customized/chosen from exists, isfile
-        and isdir.
+from pprint import pformat
+from distutils.core import Command
 
-        >>> from os.path import isdir
-        >>> isdir(find_parent_containing('.'))
+
+def find_in_parents(path, name):
+    """ Return the nearest finding in the parent dirs of path,
+       while searching for name. Returns None if no such parent dir exists.
+
+        >>> find_in_parents('/tmp/', '__not_to_be_founfd_filename.txt__')
+        >>> find_in_parents(dirname(__file__), basename(__file__)) == __file__
         True
     """
 
-    current = os.path.dirname(__file__) if path is None else path
-
-    if check == 'exists':
-        check = exists
-    elif check in ('isfile', 'file'):
-        check = isfile
-    elif check in ('isdir', 'dir'):
-        check = isdir
-
-    while not check(join(current, name)):
-        old = current
-        current = dirname(current)
-        if old == current:
+    while not isfile(join(path, name)):
+        old = path
+        path = dirname(path)
+        if old == path:
             break
     else:
-        return current
+        return join(path, name)
 
-def read_config(filename):
-    """Read the project setup.cfg file to determine versioning config."""
-    # This might raise EnvironmentError (if setup.cfg is missing), or
-    # configparser.NoSectionError (if it lacks a [versioneer] section), or
-    # configparser.NoOptionError (if it lacks "VCS="). See the docstring at
-    # the top of versioneer.py for instructions on writing your setup.cfg .
-    parser = ConfigParser()
-    parser.read(filename)
+
+def find_setup_cfg():
+    """ Return the nearest directory in the parent dirs of path,
+        that contains setup.cfg, or None if no such parent dir exists.
+
+        >>> found = find_setup_cfg()
+        >>> found is None or found.endswith('setup.cfg')
+        True
+    """
+    for path in (os.getcwd(), dirname(__file__)):
+        found = find_in_parents(path, 'setup.cfg')
+        if found:
+            return found
+
+import sys
+PYTHON = sys.version_info
+configparser_module = 'ConfigParser' if PYTHON.major == 2 else 'configparser'
+configparser = __import__(configparser_module)
+
+def parse_setup_cfg():
+    """ Find and parse the project setup.cfg that contains the versioning config.
+
+        >>> hasattr(parse_setup_cfg(), 'has_option')
+        True
+    """
+    parser = configparser.ConfigParser()
+    parser.read(find_setup_cfg())
     return parser
 
-setup_cfg = join(
-    find_parent_containing('setup.cfg', check=isfile),
-    'setup.cfg',
-)
-parser = read_config(setup_cfg)
-source_versionfile = parser.get('versioning', 'source_versionfile')
+def read_setup_cfg():
+    """ Read relevant information from setup.cfg
+
+        >>> all(x.endswith('.py') for x in read_setup_cfg())
+        True
+    """
+    section = 'versioning'
+    parser = parse_setup_cfg()
+
+    source_versionfile = parser.get(section, 'source_versionfile')
+
+    has_build = parser.has_option(section, 'build_versionfile')
+    build_versionfile = parser.get(section, 'build_versionfile') if has_build else source_versionfile
+
+    return source_versionfile, build_versionfile
+
+
 
 def import_file(name, path):
+    """ Import a python source file by its filesystem path.
+
+        >>> from os.path import join, dirname
+        >>> module = import_file('import_file_test', join(dirname(__file__), '__init__.py'))
+    """
+
+    module = None
+
+    try: # py2
+        import imp
+        module = imp.load_source(name, path)
+    except IOError:
+        pass
+
     try: # py3.5
         import importlib.util
         spec = importlib.util.spec_from_file_location(name, path)
         module = importlib.util.module_from_spec(spec)
         spec.loader.exec_module(module)
-        return module
     except:
         pass
 
     try: # py3.3, py3.4
         from importlib.machinery import SourceFileLoader
         module = SourceFileLoader(name, path).load_module()
-        return module
     except:
         pass
 
-    try: # py2
-        import imp
-        return imp.load_source(name, path)
-    except:
-        pass
-
-if __name__ != 'flowtool_versioning.dropins.cmdclass':
-
-    version_in_git = import_file('versions', source_versionfile)
-    if version_in_git:
-        get_version = version_in_git.get_version
-    else:
-        print("== Warning: source_versionfile %s could not be imported. (no tags found?)" % source_versionfile)
-        get_version = lambda: 'versionfile_not_installed'
+    return module
 
 
-def build_versionfile():
-    if parser.has_option('versioning', 'build_versionfile'):
-        return parser.get('versioning', 'build_versionfile')
-    else:
-        return source_versionfile
+def get_version():
+    """ Fallback & Test version function.
 
-#print(build_versionfile())
+        >>> get_version()
+        'no_version'
+    """
+    return 'no_version'
 
 
-def render_versionfile():
-    print(version_in_git.render_versionfile())
+def setup_versioning():
+    """ Here some magic happens.
 
-import sys
-from distutils.core import Command
+        >>> import sys
+        >>> type(setup_versioning()) in (type(sys), type(None))
+        True
+    """
+
+    global get_version
+
+    source_versionfile, build_versionfile = read_setup_cfg()
+
+    versionfile = import_file('_version', source_versionfile) if isfile(source_versionfile) else None
+
+    get_version = versionfile.get_version if hasattr(versionfile, 'get_version') else get_version
+
+    return versionfile
+
+versionfile = setup_versioning()
+
+def print_version_info():
+    """ Testable body of a setuptools/distutils command.
+
+        >>> print_version_info()
+        == Version-Config (setup.cfg):
+        ...
+    """
+    parser = parse_setup_cfg()
+
+    pretty_version_info = pformat(dict(parser.items('versioning')))
+    print('== Version-Config (setup.cfg):\n' + pretty_version_info)
+
+    if versionfile is not None and hasattr(versionfile, 'VERSION_INFO'):
+        print('== Version-Info:\n' + pformat(versionfile.VERSION_INFO))
+
 
 class cmd_version_info(Command):
+    """ Version info command.
+
+        Run `./setup.py version` to get detailed info on the latest version.
+    """
+
     description = "show versioning configuration and current project version"
     user_options = []
     boolean_options = []
 
-    def initialize_options(self):
-        pass
-
-    def finalize_options(self):
-        pass
+    def initialize_options(self): pass
+    def finalize_options(self): pass
 
     def run(self):
-        from pprint import pformat
-        print('== Version-Config (setup.cfg):\n%s' % pformat(dict(parser.items('versioning'))))
-        print('== Version-Info:\n%s' % pformat(version_in_git.VERSION_INFO))
+        print_version_info()
+
+
 
 def bump_version(info):
+    """ Bump a parsed version.
+
+        >>> bump_version({'release':(8, 1)})
+        {'release': (8, 2)}
+        >>> bump_version({'release':(8, 1), 'post_release':0})['post_release']
+        1
+        >>> bump_version({'release':(8, 1), 'pre_release':('a', 0), 'dev_release':4})['dev_release']
+        5
+        >>> bump_version({'release':(8, 1), 'pre_release':('b', 0)})['pre_release']
+        ('b', 1)
+    """
     if 'dev_release' in info:
         info['dev_release'] += 1
     elif 'post_release' in info:
@@ -162,46 +224,62 @@ def render_bumped(**kwd):
         normalized = '{}!{}'.format(kwd['epoch'], normalized)
     return normalized
 
+
+def do_bump(not_really=None):
+    """ Execute a version bump (if not testing)
+
+        >>> do_bump(not_really={'dirt':'','tag_version':{'release':(1,1,2)},'prefix':''})
+        == Next Version: {'release': (1, 1, 3)}
+        == Tagging: 1.1.3
+        >>> do_bump(not_really={'dirt':'XXX','tag_version':{'release':(1,1,2)},'prefix':''})
+        Traceback (most recent call last):
+        ...
+        SystemExit: 1
+    """
+    vcs_info = versionfile.VERSION_INFO['vcs_info'] if not_really is None else not_really
+    tag_info = bump_version(vcs_info['tag_version'])
+    print('== Next Version: %s' % pformat(tag_info))
+    if vcs_info['dirt']:
+        print("==> Auto bump aborted due to dirty git repository.")
+        sys.exit(1)
+    tag = vcs_info['prefix'] + render_bumped(**tag_info)
+    print('== Tagging: %s' % tag)
+    not_really or os.system('git tag ' + tag)
+
 class cmd_version_bump(Command):
+    """ Version bump command.
+
+        Run `./setup.py bump` to create a new git tag
+        with the smallest version component increased
+        by one.
+    """
     description = "bump the (pep440) version by adding one to the smallest version component"
     user_options = []
     boolean_options = []
 
-    def initialize_options(self):
-        pass
+    def initialize_options(self): pass
+    def finalize_options(self): pass
+    def run(self): do_bump()
 
-    def finalize_options(self):
-        pass
 
-    def run(self):
-        from pprint import pformat
-        vcs_info = version_in_git.VERSION_INFO['vcs_info']
-        tag_info = bump_version(vcs_info['tag_version'])
-        print('== Current Version:\n%s' % pformat(tag_info))
-        if vcs_info['dirt']:
-            print("==> Auto bump aborted due to dirty git repository.")
-            sys.exit(1)
-        tag = vcs_info['prefix'] + render_bumped(**tag_info)
-        print('== Tagging: %s' % tag)
-        os.system('git tag ' + tag)
+#class cmd_update_versionfile(Command):
+    #description = "update the versioning"
+    #user_options = []
+    #boolean_options = []
 
-class cmd_update_versionfile(Command):
-    description = "update the versioning"
-    user_options = []
-    boolean_options = []
+    #def initialize_options(self):
+        #pass
 
-    def initialize_options(self):
-        pass
+    #def finalize_options(self):
+        #pass
 
-    def finalize_options(self):
-        pass
+    #def run(self):
+        #print('== Updating file:\n%s' % source_versionfile)
+        #from flowtool_versioning.dropins import version
+        #versionfile = version.__file__
+        #with open(versionfile, 'r') as f_in, open(source_versionfile, 'w') as f_out:
+            #f_out.write(f_in.read())
 
-    def run(self):
-        print('== Updating file:\n%s' % source_versionfile)
-        from flowtool_versioning.dropins import version
-        versionfile = version.__file__
-        with open(versionfile, 'r') as f_in, open(source_versionfile, 'w') as f_out:
-            f_out.write(f_in.read())
 
 
 if "setuptools" in sys.modules:
@@ -209,12 +287,15 @@ if "setuptools" in sys.modules:
 else:
     from distutils.command.build_py import build_py as _build_py
 
+
 class cmd_build_py(_build_py):
-    def run(self):
-        _build_py.run(self)
+    """ It seems as if build_py is executed when the distributed package is installed. """
+
+    #def run(self):
+        #_build_py.run(self)
         # now locate _version.py in the new build/ directory and replace
         # it with an updated value
-        #deploy_to = build_versionfile()
+        #deploy_to = build_versionfile
         #print("== Deploying %s" % deploy_to)
         #deploy_versionfile(deploy_to)
 
@@ -251,24 +332,56 @@ if "setuptools" in sys.modules:
 else:
     from distutils.command.sdist import sdist as _sdist
 
+def add_to_sdist(base_dir):
+    """ The custom part of the sdist command.
+
+        >>> add_to_sdist('/tmp')
+        == Rendering:
+        ...
+        >>> add_to_sdist('/tmp')
+        == Rendering:
+        ...
+    """
+    # now locate _version.py in the new base_dir directory
+    # (remembering that it may be a hardlink) and replace it with an
+    # updated value
+
+    source_versionfile, build_versionfile = read_setup_cfg()
+    target_versionfile = os.path.join(base_dir, build_versionfile)
+    static_versionfile = versionfile.render_static_file() if versionfile else 'test_content'
+    print("== Rendering:\n%s\n== To Versionfile: %s" % (static_versionfile, target_versionfile))
+
+    try:
+        # handles the hard link case correctly
+        if os.path.exists(target_versionfile):
+            os.unlink(target_versionfile)
+        with open(target_versionfile, 'w') as fh:
+            fh.write(static_versionfile)
+    except:
+        print("=== Could not render static _version.py to sdist!")
+
+    self_target = join(base_dir, basename(__file__))
+    print("== Updating: %s" % self_target)
+    if not os.path.exists(self_target):
+        try:
+            os.link(__file__, self_target)
+        except OSError:
+            print("=== Could not add %s to sdist!" % basename(__file__))
+
+
 class cmd_sdist(_sdist):
+
     def run(self):
-        self.distribution.metadata.version = get_version()
+        self.distribution.metadata.version = setup_versioning().get_version()
         return _sdist.run(self)
 
     def make_release_tree(self, base_dir, files):
         _sdist.make_release_tree(self, base_dir, files)
-        os.link(__file__, join(base_dir, basename(__file__)))
-        # now locate _version.py in the new base_dir directory
-        # (remembering that it may be a hardlink) and replace it with an
-        # updated value
-        target_versionfile = os.path.join(base_dir, build_versionfile())
-        print("== Rendering: %s" % target_versionfile)
-        os.unlink(target_versionfile)
-        with open(target_versionfile, 'w') as fh:
-            fh.write(version_in_git.render_static_file())
+        add_to_sdist(base_dir)
+
 
 from distutils.command.upload import upload as _upload
+
 
 class cmd_upload(_upload):
 
@@ -280,23 +393,38 @@ class cmd_upload(_upload):
         print('==> For backwards compatibility you should only upload packages built with Python 3 to PyPI.')
         sys.exit(1)
 
+
+def publish_code(not_really=None):
+    """ push git and its tags
+
+        >>> publish_code(not_really=True)
+        === git push
+        === pushing tags also
+    """
+    print("=== git push")
+    not_really or os.system('git push')
+    print("=== pushing tags also")
+    not_really or os.system('git push --tags')
+
+
 class cmd_release(cmd_upload):
 
     description="Do the protected upload, but push git things first"
 
     def run(self):
-        print("=== git push")
-        os.system('git push')
-        print("=== pushing tags also")
-        os.system('git push --tags')
+        publish_code()
         return cmd_upload.run(self)
 
 
+
 def get_cmdclass():
-    """Return the custom setuptools/distutils subclasses."""
+    """ Return the custom setuptools/distutils subclasses.
+
+        >>> sorted(get_cmdclass().keys())
+        ['build_py', 'bump', 'release', 'sdist', 'upload', 'version']
+    """
     cmds = dict(
         version=cmd_version_info,
-        #versioning_update=cmd_versioning_update,
         bump=cmd_version_bump,
         build_py=cmd_build_py,
         sdist=cmd_sdist,
@@ -306,8 +434,12 @@ def get_cmdclass():
     return cmds
 
 
-def main():
-    print(get_version())
+def main(noop=None):
+    """ Prints the current version.
+
+        >>> main(True)
+    """
+    noop or print(get_version())
 
 if __name__ == '__main__':
     main()
